requirements.txt
keras
tensorflow
h5py
pillow
matplotlib
python-twitter
